[WebRTC] Add support for runtime PeeConnection setting, required after r209757 https://bugs.webkit.org/show_bug.cgi?id=166346
Patch by Alejandro G. Castro <alex@igalia.com> on 2016-12-22
Reviewed by Carlos Garcia Campos.

Source/WebKit2:

After that commit we have to enable the peerConnection runtime
flag if we want webrtc to work. We are not adding a new API for
the moment to control the peerConnection feature.

* UIProcess/API/gtk/WebKitSettings.cpp:
(webkit_settings_set_enable_media_stream): Use the mediaStream
setting to control the peerConnection

Tools:

Enable WebRTC by default to simplify testing.

* MiniBrowser/gtk/main.c:
(main):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@210096 268f45cc-cd09-0410-ab3c-d52691b4dbfc