Warning about unconnectable outlets when opening a MiniBrowser window https://bugs.webkit.org/show_bug.cgi?id=166756
Reviewed by Dan Bernstein.

* MiniBrowser/mac/BrowserWindow.xib:
When launching MiniBrowser, it says:
Failed to connect (Share) outlet from (WK1BrowserWindowController) to
    (MBToolbarItem): missing setter or instance variable
Failed to connect (shareButton) outlet from (WK1BrowserWindowController) to
    (NSButtonCell): missing setter or instance variable.

It looks like someone added the connection from the share button
to the BrowserWindowController three times; only one of the connections
is valid. Remove the other two.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@210434 268f45cc-cd09-0410-ab3c-d52691b4dbfc