Fix WinCairo build after r210319. https://bugs.webkit.org/show_bug.cgi?id=166635
* DerivedSources.cpp:
A toJS call was ambiguous, but only if the generated IndexedDB bindings are compiled all-in-one in DerivedSources.cpp.
Compiling these few files separately (which happens automatically in the CMake build) fixes the build and won't increase
the total object files beyond MSVC's current 2GB limit.
* bindings/js/JSWebGLRenderingContextBaseCustom.cpp:
(WebCore::JSWebGLRenderingContextBaseOwner::isReachableFromOpaqueRoots):
Unknown was an ambiguous symbol.  It's also defined in winioctl.h.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@210465 268f45cc-cd09-0410-ab3c-d52691b4dbfc