REGRESSION(r189555): ImageDocument title no longer includes the size of the image https://bugs.webkit.org/show_bug.cgi?id=166876 <rdar://problem/29939970>
Reviewed by Tim Horton.

Source/WebCore:

ImageDocument title no longer included the size of the image after r189555.
This is because Document::setTitle() is a no-op if the document does not
have a <head> element. To address the issue, we now include a <head>
element in ImageDocuments so that their title element properly gets added
to it.

Test: fast/images/imageDocument-title.html

* html/ImageDocument.cpp:
(WebCore::ImageDocument::createDocumentStructure):

LayoutTests:

Add layout test coverage.

* fast/images/imageDocument-title-expected.txt: Added.
* fast/images/imageDocument-title.html: Added.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@210536 268f45cc-cd09-0410-ab3c-d52691b4dbfc