Revert r210882, removing support for background-repeat-x/y https://bugs.webkit.org/show_bug.cgi?id=167223
Reviewed by Myles Maxfield.
Source/WebCore:

CSSPropertyBackgroundRepeatX and CSSPropertyBackgroundRepeatY only exist to make it easier
to parse CSSPropertyBackgroundRepeat as if it were a shorthand; they were never intended to
be web-exposed. r210882 mistakenly expose them to the web, so undo that patch, and annotate
the properties in CSSProperties.json with a flag which indicates their special status.

In addition, remove background-repeat-x/y from the list of properties allowed in video cues,
since these are not valid properties.

* css/CSSProperties.json:
* css/StyleResolver.cpp:
(WebCore::StyleResolver::isValidCueStyleProperty):
* css/makeprop.pl:
(addProperty):
* css/parser/CSSParserFastPaths.cpp:
(WebCore::CSSParserFastPaths::isKeywordPropertyID):

LayoutTests:

Modify the test to check that background-repeat-x and background-repeat-y are not parsed.

* fast/backgrounds/background-repeat-x-y-parse-expected.txt:
* fast/backgrounds/background-repeat-x-y-parse.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211149 268f45cc-cd09-0410-ab3c-d52691b4dbfc