LayoutTest media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-buttons-containers-styles.html is flaky https://bugs.webkit.org/show_bug.cgi?id=167253
Patch by Antoine Quint <graouts@apple.com> on 2017-01-25
Reviewed by Jon Lee.

Source/WebCore:

An IconButton loads its image asynchronously and waits for its load to obtain metrics. Once that happens,
its layout delegate needs to be notified that it needs a new layout. The test was a flaky failure because
the images would sometime load prior to layout, and sometime after, which would cause a failure. The test
is now robust with a controls layout happening after all buttons had loaded.

* Modules/modern-media-controls/controls/icon-button.js:
(IconButton.prototype._updateImage):
(IconButton):

LayoutTests:

This test should no longer be flaky, so let's not mark it as such anymore.

* platform/mac-wk1/TestExpectations:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211154 268f45cc-cd09-0410-ab3c-d52691b4dbfc