Several web timing tests crash in GTK+ and AppleWin bots https://bugs.webkit.org/show_bug.cgi?id=167577
Reviewed by Ryosuke Niwa.

The problem is that entry is used in both the key, to get name, and in the value with WTFMove. So, the name is
invalidated by the move. It could be fixed by simply copying the name, instead of using entry->name, but I think
that code could be simplified using HashMap::ensure and then we don't need any string copy, nor even the static
insertPerformanceEntry().

Fix crashes in several imported/w3c/web-platform-tests/user-timing/ tests.

* page/PerformanceUserTiming.cpp:
(WebCore::UserTiming::mark):
(WebCore::UserTiming::measure):
(WebCore::insertPerformanceEntry): Deleted.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211360 268f45cc-cd09-0410-ab3c-d52691b4dbfc