NULL-deref crash in TextTrack::removeCue() https://bugs.webkit.org/show_bug.cgi?id=167615
Reviewed by Eric Carlson.

It's possible for a track to be removed which was never actually added to the cue list.
Specifically, if an in-band track with a negative start or end time was parsed, it would
have been rejected by TextTrack::addCue(). When it comes time to flush those in-band cues,
TextTrack::m_cues will still be NULL. Rather than ASSERT in this case, we should revert the
behavior added in r210319 and throw an exception.

* html/track/TextTrack.cpp:
(WebCore::TextTrack::removeCue):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211401 268f45cc-cd09-0410-ab3c-d52691b4dbfc