com.apple.WebKit.WebContent.Development crashed in com.apple.WebCore: WebCore::FrameView::didDestroyRenderTree https://bugs.webkit.org/show_bug.cgi?id=167665 <rdar://problem/30241193>
Reviewed by Andreas Kling.

When destroying a PageCache entry, we would call Document::prepareFromDestruction()
which would destroy the render tree and then call FrameView::didDestroyRenderTree().
We would sometimes hit the ASSERT(m_widgetsInRenderTree.isEmpty()) assertion in
didDestroyRenderTree(). This indicates that there are still RenderWidgets associated
with the FrameView even though we claim we already destroyed the render tree.

The reason for this is that, in the PageCache context, the top document's frame is
reused on navigation and may not be associated with a new document that is currently
displayed, has a render tree (and possible RenderWidgets). The solution proposed is
to stop calling FrameView::didDestroyRenderTree() in Document::prepareFromDestruction()
if the frame's document is not |this|. This happens when the document is destroyed
while in PageCache. We do not want to mess with the frame / frameview in this case
since they are no longer associated to the present document anymore.

No new tests, already covered by existing tests which currently crash flakily.

* dom/Document.cpp:
(WebCore::Document::destroyRenderTree):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211455 268f45cc-cd09-0410-ab3c-d52691b4dbfc