WebProcess crashes in  int WTF::__throw_bad_variant_access<int> when expanding/shrinking a block selection. https://bugs.webkit.org/show_bug.cgi?id=167673 rdar://problem/30229620
Reviewed by Anders Carlsson

This is a speculative fix for a bug that might have been introduced
with http://trac.webkit.org/changeset/208479.
The code in containsRange inline fuction in WebPageIOS.mm has not been
updated when the return value of Range::compareBoundaryPoints was changed
to ExceptionOr<short>.
Since there is already a method containsRange in the Range class that
does the right thing, expandedRangeFromHandle now uses that.

* WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::expandedRangeFromHandle):
(WebKit::containsRange): Deleted.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211538 268f45cc-cd09-0410-ab3c-d52691b4dbfc