[iOS WK2] With visual viewports, a fixed bottom bar can be clipped out when the keyboard is visible https://bugs.webkit.org/show_bug.cgi?id=167710 rdar://problem/30100286
Reviewed by Wenson Hsieh.
Source/WebCore:

Add a bit of logging to show when RLC decides to not composite a fixed element which
is out of bounds.

Test: fast/visual-viewport/ios/bottom-bar-with-keyboard.html

* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::requiresCompositingForPosition):

Source/WebKit2:

The unobscuredRectInContentCoordinates passed to -didUpdateVisibleRect:... could project outside
of the bounds of the document, which is OK when rubber-banding, but not when we're in a stable state,
because that can cause fixed elements to get pushed outside the doc. This happened when the keyboard
triggered bottom content insets on the scroll view.

LayoutTests:

* fast/visual-viewport/ios/bottom-bar-with-keyboard-expected.txt: Added.
* fast/visual-viewport/ios/bottom-bar-with-keyboard.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211541 268f45cc-cd09-0410-ab3c-d52691b4dbfc