CrashTracer: [USER] com.apple.WebKit.WebContent at com.apple.WebCore: WebCore::URL::host const + 9 https://bugs.webkit.org/show_bug.cgi?id=167766 <rdar://problem/30132707>
Reviewed by Chris Dumez.

The mainframe's document pointer may be null when tearing down a page upon navigation to a page that is in the
page cache. If this triggers a GC sweep, we will attempt to reload touch bar media controls, which (as a part of
the media controller heuristic) checks the mainframe's document URL to see if quirks should be enabled. This
assumes that the mainframe's document exists, which is not a safe assumption if page navigation is occurring. As
such, we need a null check for the mainframe's document in needsPlaybackControlsManagerQuirk().

No test, as we were unable to reproduce this crash.

* html/HTMLMediaElement.cpp:
(WebCore::needsPlaybackControlsManagerQuirk):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@211613 268f45cc-cd09-0410-ab3c-d52691b4dbfc