Allow to use stored credentials also for downloads started by DownloadManager::startDownload https://bugs.webkit.org/show_bug.cgi?id=167584
Reviewed by Michael Catanzaro.

Cliking on a link that triggers a download works on an already HTTP authenticated site works, but right clicking
and downloading from the context menu shows the HTTP auth dialog again. This is because PendingDownload uses
DoNotAllowStoredCredentials unconditionally. We want to allow using cached credentials at least for
non-ephemeral sessions.

* NetworkProcess/Downloads/DownloadManager.cpp:
(WebKit::DownloadManager::startDownload):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@212291 268f45cc-cd09-0410-ab3c-d52691b4dbfc