Ensure that the end of the last invalidation point does not extend beyond the end of the buffer. https://bugs.webkit.org/show_bug.cgi?id=168786
Reviewed by Filip Pizlo.

In practice, we will always have multiple instructions after invalidation points,
and have enough room in the JIT buffer for the invalidation point to work with.
However, as a precaution, we can guarantee that there's enough room by always
emitting a label just before we link the buffer.  The label will emit nop padding
if needed.

* assembler/LinkBuffer.cpp:
(JSC::LinkBuffer::linkCode):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@212908 268f45cc-cd09-0410-ab3c-d52691b4dbfc